Catch when msgpack gem is missing and warn
When `ActiveSupport::EventReporter.encoder(:msgpack)` is called, typically during boot, we can catch when the user forgot to add the gem and give them a warning.

This is similar to how `ActiveSupport::Messages::SerializerWithFallback.[](format)` works

See:
https://github.com/rails/rails/blob/6f912a5986cd9295e3fbb45b1ec22159e51ad65d/activesupport/lib/active_support/messages/serializer_with_fallback.rb#L9-L14

Also `ActiveSupport::MessagePack`:
https://github.com/rails/rails/blob/2ca26346a563a375277e97a09d879df33682df55/activesupport/lib/active_support/message_pack.rb#L3-L10

Context:
https://github.com/rails/rails/pull/55334#discussion_r2226708177
